Workflows / CI: updated to Rails 7
Updated _dummy_ application to Rails 7.2.
Loading process has changed in Rails 7.

+++ spec/support/dummy_app.rb
@@ -50,7 +50,7 @@ class DummyApp
         yield
 
         Process.kill("KILL", out.pid)
-        File.delete("tmp/pids/server.pid")
+        File.delete("tmp/pids/server.pid") if File.exist?("tmp/pids/server.pid")
       end
     end
   end
